Edited makefile for debugging, comparse and rshell for operators
operators not complete, but made it so that tokenizer is responsible
for operators, not me. Now need to configure loops in comparse/rshell
to know what to do if string is operator. Also concerned that if I
replace semicolon with NULL, then make multiple pointers to each part of
the input that corresponds to a different command, that I'll risk losing
values.

Now realized as I typed this that this fear is bunk because the
vector<string> words in comParse isn't changed, only the vector<char*>
vector.